Extracting out banner to its own class
